+++ tests/trainer/test_trainer.py
@@ -371,8 +371,8 @@ def test_model_checkpoint_options(tmpdir, save_top_k, save_last, expected_files)
 
     # emulate callback's calls during the training
     for i, loss in enumerate(losses):
-        trainer.train_loop.current_epoch = i
-        trainer.train_loop.global_step = i
+        trainer.fit_loop.current_epoch = i
+        trainer.fit_loop.global_step = i
         trainer.logger_connector.callback_metrics.update({"checkpoint_on": loss})
         checkpoint_callback.on_validation_end(trainer, trainer.lightning_module)
 
@@ -1765,7 +1765,7 @@ def test_init_optimizers_resets_lightning_optimizers(tmpdir):
     trainer.fit(model)
     compare_optimizers()
 
-    trainer.train_loop.max_epochs = 2  # simulate multiple fit calls
+    trainer.fit_loop.max_epochs = 2  # simulate multiple fit calls
     trainer.fit(model)
     compare_optimizers()